+++ src/main/java/com/josepedevs/pcrepair/writer/PersonWriterConfig.java
@@ -12,14 +12,17 @@ import org.springframework.context.annotation.Configuration;
 @Slf4j
 @Configuration
 @RequiredArgsConstructor
-public class PersonCsvWriterConfig {
+public class PersonWriterConfig {
 
     private final PersonWriterFactory writerFactory;
 
-    @Bean("personCsvFileWriter")
-    public FlatFileItemWriter<Person> personCsvWriter(AppPropertiesReader props) {
-        log.info("Writing person values...");
-        return writerFactory.createWriter(props);
+    @Bean("personFileWriter")
+    public FlatFileItemWriter<Person> personWriter(
+            AppPropertiesReader props,
+            PersonWriterFactory writerFactory) {
+
+        final var strategy = writerFactory.getStrategy(props.getExportFormat());
+        return strategy.createWriter(props);
     }
 
 }